+++ lib/ui/wellness/WellnessSectionsContentWidget.dart
@@ -14,19 +14,18 @@
  * limitations under the License.
  */
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Transportation.dart';
-import 'package:illinois/ui/wellness/WellnessEightDimensionsPanel.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 
 class WellnessSectionsContentWidget extends StatefulWidget {
-  WellnessSectionsContentWidget();
+  final void Function()? onTapEightDimension;
+
+  WellnessSectionsContentWidget({this.onTapEightDimension});
 
   @override
   State<WellnessSectionsContentWidget> createState() => _WellnessSectionsContentWidgetState();
@@ -63,9 +62,9 @@ class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentW
   }
 
   Widget _buildContent() {
-    return Column(
+    return Padding(padding: EdgeInsets.symmetric(horizontal: 16), child: Column(
         crossAxisAlignment: CrossAxisAlignment.center,
-        children: [_buildTipDescription(), _buildEightDimensionImage(), _buildFooterDescription(), _buildEightDimensionButton()]);
+        children: [_buildTipDescription(), _buildEightDimensionImage(), _buildFooterDescription(), _buildEightDimensionButton()]));
   }
 
   Widget _buildLoadingContent() {
@@ -89,8 +88,7 @@ class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentW
   }
 
   Widget _buildEightDimensionImage() {
-    //TBD: DD - image resource for 8 dimension
-    return Container();
+    return Padding(padding: EdgeInsets.only(top: 16), child: Image.asset('images/wellness-wheel-2019.png', width: 45, height: 45));
   }
 
   Widget _buildFooterDescription() {
@@ -120,12 +118,7 @@ class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentW
         textStyle: TextStyle(fontSize: 14),
         rightIcon: Image.asset('images/external-link.png'),
         rightIconPadding: EdgeInsets.only(left: 4, right: 6),
-        onTap: _onTapEightDimensions);
-  }
-
-  void _onTapEightDimensions() {
-    Analytics().logSelect(target: "Wellness 8 Dimensions");
-    Navigator.of(context).push(CupertinoPageRoute(builder: (context) => WellnessEightDimensionsPanel()));
+        onTap: widget.onTapEightDimension ?? () => {});
   }
 
   void _setLoading(bool loading) {